Moved authentication related views to the oauth module.
